+++ api-augment/base.ts
@@ -1,3 +1,3 @@
 
-import 'https://deno.land/x/polkadot@0.2.40/rpc-augment/mod.ts';
-import 'https://deno.land/x/polkadot@0.2.40/types-augment/mod.ts';
+import 'https://deno.land/x/polkadot/rpc-augment/mod.ts';
+import 'https://deno.land/x/polkadot/types-augment/mod.ts';