+++ api-base/types/base.ts
@@ -1,6 +1,6 @@
 
 import type { Observable } from 'https://esm.sh/rxjs@7.8.1';
-import type { AnyFunction, Callback, Codec } from 'https://deno.land/x/polkadot@0.2.40/types/types/index.ts';
+import type { AnyFunction, Callback, Codec } from 'https://deno.land/x/polkadot/types/types/index.ts';
 
 export type Push<T extends readonly unknown[], V> = [...T, V]