+++ api-derive/democracy/referendumsFinished.ts
@@ -1,6 +1,6 @@
 
 import type { Observable } from 'https://esm.sh/rxjs@7.8.1';
-import type { PalletDemocracyReferendumInfo } from 'https://deno.land/x/polkadot@0.2.40/types/lookup.ts';
+import type { PalletDemocracyReferendumInfo } from 'https://deno.land/x/polkadot/types/lookup.ts';
 import type { DeriveApi } from '../types.ts';
 
 import { map, switchMap } from 'https://esm.sh/rxjs@7.8.1';