+++ api-derive/test/bytesFactory.ts
@@ -1,6 +1,6 @@
 
-import type { Bytes } from 'https://deno.land/x/polkadot@0.2.40/types/mod.ts';
-import type { Registry } from 'https://deno.land/x/polkadot@0.2.40/types/types/index.ts';
+import type { Bytes } from 'https://deno.land/x/polkadot/types/mod.ts';
+import type { Registry } from 'https://deno.land/x/polkadot/types/types/index.ts';
 
 export class BytesFactory {
   #registry: Registry;