+++ api/promise/Combinator.ts
@@ -1,8 +1,8 @@
 
-import type { Callback } from 'https://deno.land/x/polkadot@0.2.40/types/types/index.ts';
+import type { Callback } from 'https://deno.land/x/polkadot/types/types/index.ts';
 import type { UnsubscribePromise } from '../types/index.ts';
 
-import { isFunction } from 'https://deno.land/x/polkadot@0.2.40/util/mod.ts';
+import { isFunction } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 export type CombinatorCallback <T extends unknown[]> = Callback<T>;