+++ api/submittable/createClass.ts
@@ -2,16 +2,16 @@
 /* eslint-disable no-dupe-class-members */
 
 import type { Observable } from 'https://esm.sh/rxjs@7.8.1';
-import type { Address, ApplyExtrinsicResult, Call, Extrinsic, ExtrinsicEra, ExtrinsicStatus, Hash, Header, Index, RuntimeDispatchInfo, SignerPayload } from 'https://deno.land/x/polkadot@0.2.40/types/interfaces/index.ts';
-import type { Callback, Codec, Constructor, ISubmittableResult, SignatureOptions } from 'https://deno.land/x/polkadot@0.2.40/types/types/index.ts';
-import type { Registry } from 'https://deno.land/x/polkadot@0.2.40/types-codec/types/index.ts';
+import type { Address, ApplyExtrinsicResult, Call, Extrinsic, ExtrinsicEra, ExtrinsicStatus, Hash, Header, Index, RuntimeDispatchInfo, SignerPayload } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { Callback, Codec, Constructor, ISubmittableResult, SignatureOptions } from 'https://deno.land/x/polkadot/types/types/index.ts';
+import type { Registry } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
 import type { ApiBase } from '../base/index.ts';
 import type { ApiInterfaceRx, ApiTypes, PromiseOrObs, SignerResult } from '../types/index.ts';
 import type { AddressOrPair, SignerOptions, SubmittableDryRunResult, SubmittableExtrinsic, SubmittablePaymentResult, SubmittableResultResult, SubmittableResultSubscription } from './types.ts';
 
 import { catchError, first, map, mergeMap, of, switchMap, tap } from 'https://esm.sh/rxjs@7.8.1';
 
-import { identity, isBn, isFunction, isNumber, isString, isU8a, objectSpread } from 'https://deno.land/x/polkadot@0.2.40/util/mod.ts';
+import { identity, isBn, isFunction, isNumber, isString, isU8a, objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { filterEvents, isKeyringPair } from '../util/index.ts';
 import { SubmittableResult } from './Result.ts';
@@ -158,9 +158,11 @@ export function createClass <ApiType extends ApiTypes> ({ api, apiType, blockHas
               // setup our options (same way as in signAndSend)
               const eraOptions = makeEraOptions(api, this.registry, allOptions, signingInfo);
               const signOptions = makeSignOptions(api, eraOptions, {});
-              const u8a = this.isSigned
-                ? api.tx(this).signFake(address, signOptions).toU8a()
-                : this.signFake(address, signOptions).toU8a();
+
+              // 1. Don't use the internal objects inside the new tx (hence toU8a)
+              // 2. Don't override the data from existing signed extrinsics
+              // 3. Ensure that this object stays intact, with no new sign after operation
+              const u8a = api.tx(this.toU8a()).signFake(address, signOptions).toU8a();
 
               return api.call.transactionPaymentApi.queryInfo(u8a, u8a.length);
             })