+++ rpc-provider/substrate-connect/Health.ts
@@ -1,7 +1,7 @@
 
 import type { HealthChecker, SmoldotHealth } from './types.ts';
 
-import { stringify } from 'https://deno.land/x/polkadot@0.2.40/util/mod.ts';
+import { stringify } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 interface JSONRequest {
   id: string;